requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sentence-transformers
chromadb
serpapi
scikit-learn
requests
beautifulsoup4
easyocr
opencv-python
